fix: move DiracTime to DiracCommon
